Remove tex_coord range checks
This was causing multiple bugs in texture rendering. The crab model
would only show half of the texture since it was reversed, and the cube
texture had empty lines in it in orthographic mode.

This was only a temporary fix until we had screen clipping. We still
don't have it but I want to get rid of the bug now before I forget.